feat: 데이터 주입 기능 구현

+++ src/main/java/com/ohgiraffers/adregamdi/place/command/domain/aggregate/vo/CategoryVO.java
@@ -14,6 +14,6 @@ import javax.persistence.Embeddable;
 public class CategoryVO {
 
     @Column(nullable = false, name = "CATEGORY_NO")
-    private Long categoryNo;
+    private int categoryNo;
 
 }
